backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.22
alembic==1.12.0
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-dotenv==1.0.0
ccxt==4.4.74
pandas==2.1.1
numpy==1.26.0
requests==2.31.0
openai==1.2.0
anthropic==0.5.0
google-generativeai==0.3.0
deepseek==1.0.0
feedparser==6.0.10
email-validator
web3
eth-account
